refactor: 피드 추상화 작업 (#63)
