+++ app/(auth)/src/ui/dialog/delete-dialog.tsx
@@ -1,0 +1,70 @@
+import { ReactElement } from "react"
+
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog"
+import { buttonVariants } from "@/components/ui/button"
+
+interface DeleteDialogProps {
+  children: ReactElement
+  title: string
+  description: string
+  onDelete: () => void
+}
+
+const DeleteDialog = ({
+  children,
+  title,
+  description,
+  onDelete,
+}: DeleteDialogProps) => {
+  return (
+    <AlertDialog>
+      <AlertDialogTrigger
+        asChild
+        onClick={(e) => {
+          e.stopPropagation()
+        }}
+      >
+        {children}
+      </AlertDialogTrigger>
+
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle className="select-none">{title}</AlertDialogTitle>
+
+          <AlertDialogDescription className="select-none">
+            {description}
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+
+        <AlertDialogFooter>
+          <AlertDialogCancel onClick={(e) => e.stopPropagation()}>
+            취소
+          </AlertDialogCancel>
+          <AlertDialogAction
+            onClick={(e) => {
+              e.stopPropagation()
+              onDelete()
+            }}
+            className={buttonVariants({
+              variant: "destructive",
+            })}
+          >
+            삭제
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  )
+}
+
+export default DeleteDialog
