+++ app/(auth)/src/ui/sheet/comment-sheet.tsx
@@ -3,7 +3,7 @@
 import { useMemo } from "react"
 import { useRouter, useSearchParams } from "next/navigation"
 
-import { MonsterList, monsterList } from "@/interface/monster"
+import { MonsterList } from "@/interface/monster"
 import { userRole } from "@/interface/user"
 import clsx from "clsx"
 import {
@@ -14,21 +14,8 @@ import {
 } from "framer-motion"
 import { useSession } from "next-auth/react"
 
-import { dateDistanceToNow } from "@/lib/date"
 import { apiErrorMessage } from "@/lib/error-message"
 import { pageRoute } from "@/lib/page-route"
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from "@/components/ui/alert-dialog"
-import { Button, buttonVariants } from "@/components/ui/button"
 import {
   Sheet,
   SheetContent,
@@ -41,6 +28,7 @@ import MonsterImage from "@/components/common/monster-image"
 
 import useCommentDelete from "../../hooks/useCommentDelete"
 import useCommentList from "../../hooks/useCommentList"
+import CommentCard from "../card/comment-card"
 import CommentDialog from "../dialog/comment-dialog"
 
 const animations: AnimationProps = {
@@ -68,7 +56,7 @@ export default function CommentSheet() {
 
   const isLoading = isCommentListLoading || isDeleteCommentLoading
 
-  const userId = session?.user.id
+  const sessionUserId = session?.user.id
 
   const isAdmin = session?.user.role === userRole.Enum.ADMIN
 
@@ -167,121 +155,29 @@ export default function CommentSheet() {
               {hasCommentList && (
                 <ul className="grid grid-cols-2 gap-4 max-lg:grid-cols-1">
                   <AnimatePresence>
-                    {commentList?.map(
-                      ({
-                        id,
-                        author,
-                        createdAt,
-                        monsterList: attackMonsterList,
-                      }) => {
-                        const commentList = monsterList.parse(attackMonsterList)
-
-                        return (
-                          <motion.li
-                            key={id}
-                            layout
-                            style={
-                              isPresent
-                                ? { position: "static" }
-                                : { position: "absolute" }
+                    {commentList?.map((props) => {
+                      return (
+                        <motion.li
+                          key={props.id}
+                          layout
+                          style={
+                            isPresent
+                              ? { position: "static" }
+                              : { position: "absolute" }
+                          }
+                          {...animations}
+                        >
+                          <CommentCard
+                            {...props}
+                            onDelete={
+                              isAdmin || sessionUserId === props.author.id
+                                ? (id) => handleCommentDelete(id)
+                                : undefined
                             }
-                            {...animations}
-                          >
-                            <div className="group relative flex h-max cursor-pointer justify-between gap-6 overflow-hidden rounded-lg border p-4 shadow-sm transition-shadow duration-300 hover:shadow-md">
-                              <div className="flex flex-col items-center gap-2 backdrop-blur-sm">
-                                <div className="flex flex-col gap-1">
-                                  <span className="flex select-none flex-col">
-                                    <em className="flex items-center text-xs font-semibold not-italic text-gray-400">
-                                      {author.role === userRole.Enum.ADMIN ? (
-                                        <Icons.crown
-                                          className="mr-0.5"
-                                          size={14}
-                                        />
-                                      ) : (
-                                        <Icons.award size={14} />
-                                      )}
-                                      {author.guildName}
-                                    </em>
-                                    <em className="text-sm font-semibold not-italic text-foreground">
-                                      {author.nickname}
-                                    </em>
-                                  </span>
-
-                                  <span className="flex select-none gap-1 text-xs font-medium text-gray-400">
-                                    {dateDistanceToNow(createdAt)}
-                                  </span>
-                                </div>
-                              </div>
-
-                              <div className="relative z-50 flex h-4/6 items-center justify-center gap-2">
-                                {commentList.map((monsterInfo) => {
-                                  return (
-                                    <MonsterImage
-                                      key={monsterInfo.id}
-                                      className="cursor-auto drop-shadow-lg"
-                                      monsterInfo={monsterInfo}
-                                    />
-                                  )
-                                })}
-                              </div>
-
-                              <div
-                                className={clsx(
-                                  "absolute inset-0 z-50 flex items-center justify-center opacity-0 transition-all bg-foreground/30",
-                                  {
-                                    "group-hover:opacity-100":
-                                      isAdmin || userId === author.id,
-                                  }
-                                )}
-                              >
-                                <AlertDialog>
-                                  <AlertDialogTrigger asChild>
-                                    <Button
-                                      className="hover:bg-transparent"
-                                      variant="ghost"
-                                      size="menu"
-                                    >
-                                      <i className="global-menu remove" />
-                                    </Button>
-                                  </AlertDialogTrigger>
-
-                                  <AlertDialogContent>
-                                    <AlertDialogHeader>
-                                      <AlertDialogTitle className="select-none">
-                                        공격덱을 삭제하시겠습니까?
-                                      </AlertDialogTitle>
-
-                                      <AlertDialogDescription className="select-none">
-                                        삭제한 공격덱은 복구가 불가능합니다.
-                                      </AlertDialogDescription>
-                                    </AlertDialogHeader>
-
-                                    <AlertDialogFooter>
-                                      <AlertDialogCancel>
-                                        취소
-                                      </AlertDialogCancel>
-                                      <AlertDialogAction
-                                        disabled={!feedId}
-                                        onClick={(e) => {
-                                          e.stopPropagation()
-
-                                          handleCommentDelete(id)
-                                        }}
-                                        className={buttonVariants({
-                                          variant: "destructive",
-                                        })}
-                                      >
-                                        삭제
-                                      </AlertDialogAction>
-                                    </AlertDialogFooter>
-                                  </AlertDialogContent>
-                                </AlertDialog>
-                              </div>
-                            </div>
-                          </motion.li>
-                        )
-                      }
-                    )}
+                          />
+                        </motion.li>
+                      )
+                    })}
                   </AnimatePresence>
                 </ul>
               )}
